requirements.txt
python-chess==1.999
stockfish==3.28.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
streamlit>=1.28.0
plotly>=5.17.0
pyinstaller>=6.3.0
requests>=2.31.0

